Re-structure the Probability Encoder to Belief Propagation
