+++ microbat/src/main/microbat/handler/BaselineHandler.java
@@ -13,7 +13,7 @@ import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.swt.widgets.Display;
 
 import microbat.baseline.encoders.NodeFeedbackPair;
-import microbat.baseline.encoders.ProbabilityEncoder;
+import microbat.baseline.encoders.BeliefPropagation;
 import microbat.model.trace.Trace;
 import microbat.model.trace.TraceNode;
 import microbat.model.value.VarValue;
@@ -53,7 +53,7 @@ public class BaselineHandler extends AbstractHandler {
 				Trace trace = traceView.getTrace();
 				
 				// Setup the probability encoder
-				ProbabilityEncoder encoder = new ProbabilityEncoder(trace);
+				BeliefPropagation encoder = new BeliefPropagation(trace);
 				encoder.setInputVars(BaselineHandler.inputs);
 				encoder.setOutputVars(BaselineHandler.outputs);
 				encoder.setup();
@@ -93,7 +93,7 @@ public class BaselineHandler extends AbstractHandler {
 					NodeFeedbackPair nodeFeedbackPair = new NodeFeedbackPair(feedbackNode, feedback);
 					BaselineHandler.resetManualFeedback();
 					
-					ProbabilityEncoder.addFeedback(nodeFeedbackPair);;
+					BeliefPropagation.addFeedback(nodeFeedbackPair);;
 					noOfFeedbacks += 1;
 				}
