+++ microbat/src/test/microbat/baseline/constraints/MessageProcessorTest.java
@@ -9,7 +9,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import microbat.baseline.encoders.ProbabilityEncoder;
+import microbat.baseline.encoders.BeliefPropagation;
 import microbat.baseline.factorgraph.MessageProcessor;
 import microbat.baseline.factorgraph.VarIDConverter;
 import microbat.model.trace.TraceNode;
@@ -62,9 +62,9 @@ public class MessageProcessorTest {
 		this.controlDom = new TraceNode(null, null, 1, null, "");
 		
 		final String type = "boolean";
-		final String varName = ProbabilityEncoder.CONDITION_RESULT_NAME_PRE + this.controlDom.getOrder();
+		final String varName = BeliefPropagation.CONDITION_RESULT_NAME_PRE + this.controlDom.getOrder();
 		
-		this.controlDomValueID = ProbabilityEncoder.CONDITION_RESULT_ID_PRE + this.controlDom.getOrder();
+		this.controlDomValueID = BeliefPropagation.CONDITION_RESULT_ID_PRE + this.controlDom.getOrder();
 		
 		Variable variable = new LocalVar(varName, type, "", 1);
 		VarValue conditionResult = new PrimitiveValue("1", true, variable);
