+++ microbat/src/test/microbat/baseline/constraints/VariableConstraintA1Test.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
 
-import microbat.baseline.encoders.ProbabilityEncoder;
+import microbat.baseline.encoders.BeliefPropagation;
 import microbat.model.trace.TraceNode;
 import microbat.model.value.PrimitiveValue;
 import microbat.model.value.VarValue;
@@ -52,9 +52,9 @@ public class VariableConstraintA1Test {
 		this.controlDom = new TraceNode(null, null, 1, null, "");
 		
 		final String type = "boolean";
-		final String varName = ProbabilityEncoder.CONDITION_RESULT_NAME_PRE + this.controlDom.getOrder();
+		final String varName = BeliefPropagation.CONDITION_RESULT_NAME_PRE + this.controlDom.getOrder();
 		
-		this.controlDomValueID = ProbabilityEncoder.CONDITION_RESULT_ID_PRE + this.controlDom.getOrder();
+		this.controlDomValueID = BeliefPropagation.CONDITION_RESULT_ID_PRE + this.controlDom.getOrder();
 		
 		Variable variable = new LocalVar(varName, type, "", 1);
 		VarValue conditionResult = new PrimitiveValue("1", true, variable);
